teo_refresh_cos/requirements.txt
tencentcloud-sdk-python-teo==3.0.1256
